Refactor namespaces for repository organization
This commit refactors the namespace from
`DigitalRightsManagement.Infrastructure.Persistence` to
`DigitalRightsManagement.Infrastructure.Persistence.Repositories`
across multiple files. This restructuring aims to better
organize repository-related classes and interfaces, enhancing
maintainability and clarity in the architecture.

Import statements have been updated accordingly to ensure
the code functions correctly post-refactor. Affected files
include dependency injection configurations, database
management interfaces, entity framework migrations, and
repository implementations, indicating a comprehensive
update to the repository pattern within the application.
