+++ src/DigitalRightsManagement.Infrastructure/Persistence/PersistenceDependencyInjection.cs
@@ -1,5 +1,6 @@
 ﻿using DigitalRightsManagement.Application.Persistence;
 using DigitalRightsManagement.Common;
+using DigitalRightsManagement.Infrastructure.Persistence.Repositories;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
