+++ src/DigitalRightsManagement.Infrastructure/Persistence/Repositories/ApplicationDbContext.cs
@@ -6,7 +6,7 @@ using DigitalRightsManagement.Infrastructure.Persistence.EntityTypeConfiguration
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace DigitalRightsManagement.Infrastructure.Persistence;
+namespace DigitalRightsManagement.Infrastructure.Persistence.Repositories;
 
 internal sealed class ApplicationDbContext(IPublisher publisher, DbContextOptions<ApplicationDbContext> options) : DbContext(options), IUnitOfWork
 {
